test(runtime-core): moved tests to correct test file
