+++ test-dts/component.test-d.ts
@@ -11,9 +11,7 @@ import {
   FunctionalComponent,
   ComponentPublicInstance,
   toRefs,
-  IsAny,
-  getCurrentInstance,
-  ComponentCustomProperties
+  IsAny
 } from './index'
 
 declare function extractComponentOptions<Props, RawBindings>(
@@ -477,13 +475,4 @@ describe('class', () => {
   const { props } = extractComponentOptions(MyComponent)
 
   expectType<number>(props.foo)
-})
-
-describe('getCurrentInstance', () => {
-  const instance = getCurrentInstance()!
-
-  expectType<ComponentCustomProperties>(
-    instance.appContext.config.globalProperties
-  )
-  expectType<ComponentCustomProperties>(instance.proxy!)
 })
